requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.5
colorgram.py==1.2.0
colorlog==6.6.0
colormap==1.0.4
convcolors==2.2.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Django==3.1.3
easydev==0.12.0
entrypoints==0.4
executing==0.8.3
extcolors==1.0.0
fonttools==4.33.3
gunicorn==20.1.0
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.2.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tornado==6.1
traitlets==5.3.0
tzdata==2022.1
wcwidth==0.2.5
whitenoise==6.2.0
